result dto generic하게 수정 및 id int로 수정

+++ src/main/kotlin/com/ojajae/domain/dashbord/repository/DashboardRepository.kt
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport
 import org.springframework.stereotype.Repository
 
-interface DashboardRepository: JpaRepository<Dashboard, Long> {
+interface DashboardRepository: JpaRepository<Dashboard, Int> {
 }
 
 @Repository
