+++ src/main/kotlin/com/ojajae/domain/item_tag/entity/ItemTag.kt
@@ -8,4 +8,4 @@ class ItemTag(
     var name: String,
 
     var imageUrl: String,
-): MutableEntity<Long>()
+): MutableEntity<Int>()
