+++ src/main/kotlin/com/ojajae/domain/store/entity/Store.kt
@@ -26,4 +26,4 @@ class Store(
     var identificationNumber: String?,
 
     var items: String?,
-): MutableEntity<Long>()
+): MutableEntity<Int>()
