+++ src/main/kotlin/com/ojajae/domain/store/entity/StoreFile.kt
@@ -10,4 +10,4 @@ class StoreFile(
     var fileUrl: String,
 
     var order: Int = 0,
-) : ImmutableEntity<Long>()
+) : ImmutableEntity<Int>()
