+++ src/main/kotlin/com/ojajae/domain/store/form/request/StoreRequestForm.kt
@@ -1,5 +1,5 @@
 package com.ojajae.domain.store.form.request
 
 data class StoreRequestForm(
-    val storeId: Long
+    val storeId: Int
 )
